Fixed the error message of the encryption function
I lied on the last commit. I made the program check if the characters the user typed, were the characters the user typed. Whatever the user typed would be stored in a list (message). I made the program loop through each character of the list to check if it was in the list.