Adding new cipher
Before working on the decrypting function I realized it'll do the exact same thing as the encrypting function. The encrypting function reverses the users input and prints it out. The decrypting function will do the same, reverse the users input and print it out. Instead of wasting code by making two functions that do the same thing I decided to make the program able to work with two ciphers. I renamed the encrypting function to atbash. I will be adding the affine cipher to the program.