Fixed an error for an error, fixed an error, and removed unnecessary pseudocode.
Originally, when you type a message that has characters that the cipher didn't have, the program would print out an error message. But since the error message was in a loop, it would be printed as many times as there are characters in the message. Instead of asking the user for a message in the atbash function, I put it in the choice function along with the else statement.

I noticed another problem. When I want to restart the program, the final_message list kept it's value instead of reseting to an empty list. I moved it inside the atbash function to fix it.

Removed unnecessary pseudocode